refactor: added user and item model
