+++ section6/code/resources/item.py
@@ -2,6 +2,8 @@ from flask_restful import Resource, reqparse
 from flask_jwt import jwt_required
 import sqlite3
 
+from models.item import ItemModel
+
 
 class Item(Resource):
     parser = reqparse.RequestParser()
@@ -12,77 +14,44 @@ class Item(Resource):
         help="This field cannot be left blank!"
     )
 
-    @classmethod
-    def insert(cls, item):
-        connection = sqlite3.connect('data.db')
-        cursor = connection.cursor()
-
-        query = "INSERT INTO items VALUES (?, ?)"
-        cursor.execute(query, (item['name'], item['price']))
-        connection.commit()
-        connection.close()
-
-    @classmethod
-    def update(cls, item):
-        connection = sqlite3.connect('data.db')
-        cursor = connection.cursor()
-
-        query = "UPDATE items SET price = ? WHERE name = ?"
-        cursor.execute(query, (item['price'], item['name']))
-        connection.commit()
-        connection.close()
-
-    @classmethod
-    def find_by_name(cls, name):
-        connection = sqlite3.connect('data.db')
-        cursor = connection.cursor()
-
-        query = "SELECT * FROM items WHERE name = ?"
-        result = cursor.execute(query, (name,))
-        row = result.fetchone()
-        connection.close()
-
-        if row:
-            return {'item': {'name': row[0], 'price': row[1]}}
-
     @jwt_required()
     def get(self, name):
-        item = Item.find_by_name(name)
-        return item if item else {'message': 'Item not found'}, 404
+        item = ItemModel.find_by_name(name)
+        return item.json() if item else {'message': 'Item not found'}, 404
 
     def post(self, name):
-        if Item.find_by_name(name):
+        if ItemModel.find_by_name(name):
             return {'message', "An item with name '{}' already exists.".format(name)}, 400
 
         data = Item.parser.parse_args()
 
-        item = {'name': name, 'price': data['price']}
+        item = ItemModel(name, data['price'])
 
         try:
-            Item.insert(item)
+            item.insert()
         except:
             return {'message': 'An error occurred inserting the item.'}, 500
 
-        return item, 201
+        return item.json(), 201
 
     def put(self, name):
         data = Item.parser.parse_args()
 
-        item = Item.find_by_name(name)
-        updated_item = {'item': {'name': name, 'price': data['price']}}
+        item = ItemModel.find_by_name(name)
+        updated_item = ItemModel(name, data['price'])
 
         if item is None:
             try:
-                Item.insert(updated_item)
+                updated_item.insert()
             except:
                 return {'message': 'An error occurred inserting the item'}, 500
         else:
             try:
-                Item.update(updated_item)
+                updated_item.update()
             except:
                 return {'message': 'An error occurred updating the item'}, 500
 
-        return updated_item
+        return updated_item.json()
 
     def delete(self, name):
         connection = sqlite3.connect('data.db')
