+++ section6/code/resources/user.py
@@ -1,0 +1,36 @@
+import sqlite3
+from models.user import UserModel
+from flask_restful import Resource, reqparse
+
+
+class UserRegister(Resource):
+    parser = reqparse.RequestParser()
+    parser.add_argument(
+        'username',
+        type=str,
+        required=True,
+        help="This field cannot be left blank!"
+    )
+    parser.add_argument(
+        'password',
+        type=str,
+        required=True,
+        help="This field cannot be left blank!"
+    )
+
+    def post(self):
+        data = UserRegister.parser.parse_args()
+
+        connection = sqlite3.connect('data.db')
+        cursor = connection.cursor()
+
+        if UserModel.find_by_username(data['username']):
+            return {"message": "A user with username {} already exists.".format(data['username'])}, 400
+
+        query = "INSERT INTO users VALUES (NULL, ?, ?)"
+        cursor.execute(query, (data['username'], data['password']))
+
+        connection.commit()
+        connection.close()
+
+        return {"message": "User created successfully."}, 201
